pyproject.toml
[tool.poetry]
name = "idn-github"
version = "0.1.0"
description = ""
authors = ["jamiefeiss <jamie.feiss@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
jwt = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
